Code/GraphMol/Descriptors/Wrap/rdMolDescriptors.cpp Code/GraphMol/ForceFieldHelpers/MMFF/MMFF.h

+++ Code/GraphMol/ForceFieldHelpers/MMFF/MMFF.h
@@ -94,7 +94,7 @@ inline void MMFFOptimizeMoleculeConfs(ROMol &mol,
   } else {
     res.resize(mol.getNumConformers());
     for (unsigned int i = 0; i < mol.getNumConformers(); ++i) {
-      res[i] = std::make_pair(static_cast<int>(-1), static_cast<double>(-1));
+      res[i] = std::make_pair((-1), static_cast<double>(-1));
     }
   }
 }
